feat: 1282 refacto send mail
